% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sticky.R
\name{draw.stick}
\alias{draw.stick}
\title{Draw a stick figure}
\usage{
draw.stick(x, y, scale = 1, arms = "down", gender = "male", lwd = 3,
  clcol = "white", face = "happy", linecol = gray(0.3), hat = NA)
}
\description{
This code draws a stick figure in an existing plot device
}
\seealso{
\url{https://github.com/EconometricsBySimulation/R-Graphics/blob/master/Stick-Figures/draw.stick.R}
}

