% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{hi}
\alias{hi}
\title{Help start conversations}
\usage{
hi(who, how = c("shout", "whisper"))
}
\arguments{
\item{who}{character(1) The name of the person you wish to talk to.}

\item{how}{character(1) How to greet the conversant}
}
\value{
character(1) A line of text to be used when statrting conversation.
}
\description{
This function generates some helpful text for causal conversations
}
\examples{
hi("Ziqiang Chen")
hi("Ziqiang Chen", "whisper")

}

