% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.R
\name{input_classic}
\alias{input_classic}
\title{Classic R approaches to data analysis}
\usage{
input_classic(pdata_file, exprs_file)
}
\arguments{
\item{pdata_file}{character(1) The path to the phenotype data file.}

\item{exprs_file}{character(1) The path to the expression data file.}
}
\value{
A data.frame() containing samples as rows and phenotype data
     and expression data as columns
}
\description{
The function here illustrate how a 'old-school' R user might approach
data management and analysis.
}
\details{
There are two files.
}

