% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich.R
\name{input_rich}
\alias{input_rich}
\title{Rich data input}
\usage{
input_rich(pdata_file, exprs_file)
}
\arguments{
\item{pdata_file}{character(1) The path to the phenotype data file.}

\item{exprs_file}{character(1) The path to the expression data file.}
}
\value{
A SummarizedExperiment() containing samples as colData(),
     and expression values as assay().
}
\description{
Capture the semantic meaning of data that we present.
For instance, separately recongnizing...
}

