% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_scalar_character}
\alias{is_scalar_character}
\title{Test that argument is a scalar character}
\usage{
is_scalar_character(x)
}
\arguments{
\item{x}{character(1)}
}
\value{
logical(1)
}
\description{
Test that argument is a scalar character
}

