% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{hi}
\alias{hi}
\title{Help start conversations}
\usage{
hi(who)
}
\arguments{
\item{who}{character(1) The name of the person who you wish to start
a conversation with,}
}
\value{
character(1) A kine if text to be used when starting conversations.
}
\description{
This function ...
}
\examples{
 hi("Song Liu")


}

