% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{hi}
\alias{hi}
\title{Help start conversations}
\usage{
hi(who, how = c("shout", "whisper"))
}
\arguments{
\item{who}{character(1) The name of the person who you wish to start
a conversation with,}

\item{how}{character(1) How to greet the conversant. Either "shout" (default)
or "whisper"}
}
\value{
character(1) A kine if text to be used when starting conversations.
}
\description{
This function ...
}
\examples{
 hi("Song Liu")


}

