% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{hi}
\alias{hi}
\title{Help start conversations}
\usage{
hi(who, how = c("shout", "whisper"))
}
\arguments{
\item{who}{character(1) The name of the person you wish to start a
conversation with.}

\item{how}{character(1) How to greet the conversant. Either "shout" (default)
or "whisper".}
}
\value{
character(1) A line of text to be used when starting conversations.
}
\description{
This function generates some helpful text that can be used to start
conversations in all kinds of awkward social situations.
}
\examples{
hi("Martin Morgan")
hi("Martin Morgan", "whisper")

}

