% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{hi}
\alias{hi}
\title{help starts}
\usage{
hi(who)
}
\arguments{
\item{who}{character(1) the name of person to start a conversation with}
}
\value{
character() a line of text to be used when starting conversation.
}
\description{
This function returns numbers of letters
}
\examples{
hi("world")

}

