% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{input_classic}
\alias{input_classic}
\title{Expression class}
\usage{
input_classic(pdata_file, exprs_file)
}
\arguments{
\item{pdata_file}{phenotype file}

\item{exprs_file}{gene exprssion data file}
}
\value{
A data.fram
}
\description{
blabla...
}

