% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.R
\name{input_classic}
\alias{input_classic}
\title{Classic R approaches to data analysis}
\usage{
input_classic(pdata_file, exprs_file)
}
\arguments{
\item{pdata_file}{character(1) The path to the pheno data file.}

\item{exprs_file}{character(1) The path to the expression file.}
}
\value{
A data.frame containing samples as rows and phenotype data and
expression data as columns.
}
\description{
The functions here ... data management and analysis
}
\details{
There are two files. They are from a micorarray experiments.
The 1st file, All-phenotype.csv descrbes the 128 samples. The 2nd file,
is the exeriment data file
}

