% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.R
\name{input_classic}
\alias{input_classic}
\title{Classic R approaches}
\usage{
input_classic(pdata_file, exprs_file)
}
\arguments{
\item{pdata_file}{character(1)}

\item{exprs_file}{character(1)}
}
\value{
A data.frame() containing samples as rows and phenotype data and
expression data as columns.
}
\description{
Classic R approaches
}
\details{
There are two files, from a microarray experiment. ALL-phenoData.csv
describes the 128 samples, ALL-expression.csv are the normalized expressions
values for 12k probesets across 128 samples.
}

