% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich.R
\name{input_rich}
\alias{input_rich}
\title{Rich data input}
\usage{
input_rich(pdata_file, expr_file)
}
\arguments{
\item{pdata_file}{charactger (1) The path to the pheno data
file}

\item{exprs_file}{character(1) The path to the expression file}
}
\value{
A SummarizedExperiment() containing samples as
      coldata() and features as rows
}
\description{
Capture the semantic meaning of the data that
we represent. For instance, separately
recognizing the 'phenotypic' data (describing
the samples ) and expression values
}

