% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rish.R
\name{input_rich}
\alias{input_rich}
\title{Rich data input}
\usage{
input_rich(pdata_file, exprs_file)
}
\arguments{
\item{pdata_file}{phenotype file}

\item{exprs_file}{gene exprssion data file}
}
\value{
A SummarizedExperiment() containing samples as colData(), and expression values
   as assay().
}
\description{
Caputre the semantic meaing. For instance, separately recognizing the 'phenotypic' data
}
\examples{
   #se <- input_rich(pdata_file, exprs_file)
   #se[,se$mol.biol \%in\% c("BCR/ABL", "NEG")]
   #assays(se)[[1]] \%>\% rowMeans() \%>\% density() \%>\% plot()
}

