% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich.R
\name{input_rich}
\alias{input_rich}
\title{Rich data inout}
\usage{
input_rich(pdata_file, exprs_file)
}
\arguments{
\item{pdata_file}{phenotype file}

\item{exprs_file}{gene expression data file}
}
\value{
A SummariedExperiment() containing samples as colData(), and the
expression value as assay().
}
\description{
Capture the semantic meaning of the data that
we represent. For instance, separately
recognizing the 'phenotypic' data (descrbing
the samples) and the 'expression' values
(derived from, e.g., microarray or RNAseq
experiment).
}

