% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich.R
\name{input_rich}
\alias{input_rich}
\title{Rich data input}
\usage{
input_rich(pdata_file, exprs_file)
}
\arguments{
\item{pdata_file}{character(1) The path to the pheno data
file.}

\item{exprs_file}{character(1) The path to the expression
file.}
}
\value{
A SummarizedExperiment() containing samples as
    colData(), and the expression values as assay().
}
\description{
Capture the semantic meaning of the data that
we represent. For instance, separately
recognizing the 'phenotypic' data (descrbing
the samples) and the 'expression' values 
(derived from, e.g., microarray or RNAseq
experiment).
}

