% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{input_tidy}
\alias{input_tidy}
\title{Represent data in a 'tidy' format}
\usage{
input_tidy(pdata_file, exprs_file)
}
\arguments{
\item{pdata_file}{phenotype file}

\item{exprs_file}{gene exprssion data file}
}
\value{
A data.frame()
}
\description{
'Tidy' data takes a 'long' view of data
}
\details{
see more from 'dplyr'
}

